requirements.txt
Flask==2.0.1
pymongo==3.12.0
Flask-PyMongo==2.3.0
bcrypt==3.2.0
python-dotenv==0.19.0
Flask-Login==0.5.0 